+++ src/packages/core/src/application/theme.tsx
@@ -10,24 +10,6 @@ export const buildEverypageTheme = (pageTheme: RecursivePartial<ITheme>): ITheme
       paddingSectionTop: '5em',
       paddingSectionBottom: '5em',
     },
-    links: {
-      footerLink: {
-        normal: {
-          default: {
-            text: {
-              color: '$colors.brandPrimary',
-              'text-decoration': 'none',
-            },
-          },
-          hover: {
-            text: {
-              'text-decoration': 'underline',
-            },
-          },
-        },
-
-      },
-    },
     boxes: {
       iconHolder: {
         background: '$colors.brandPrimary',