+++ src/packages/core/src/sections/contact2.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { getClassName } from '@kibalabs/core';
-import { Alignment, Button, Direction, Form, InputType, KibaIcon, MultiLineInput, PaddingSize, ResponsiveContainingView, ResponsiveTextAlignmentView, SingleLineInput, Spacing, Stack, Text, TextAlignment } from '@kibalabs/ui-react';
+import { Alignment, Button, Direction, Form, InputType, KibaIcon, MultiLineInput, PaddingSize, ResponsiveContainingView, ResponsiveTextAlignmentView, SingleLineInput, Spacing, Stack, Text, TextAlignment, MarkdownText } from '@kibalabs/ui-react';
 
 import { ISectionProps, Section } from '.';
 import { SectionSubtitleText, SectionTitleText } from '../components';
@@ -12,11 +12,10 @@ import { IFormProps } from '../model';
 interface IContact2Props extends ISectionProps, IFormProps {
   titleText?: string;
   subtitleText?: string;
-  addressLine1: string;
-  addressLine2: string;
-  countryCode: number;
-  phoneNumber: number;
-  email: string;
+  bottomText?: string;
+  address?: string;
+  phoneNumber?: number;
+  email?: string;
   // TODO(krishan711): think about how the other input fields should be customized.
   inputButtonText?: string;
   inputSuccessMessageText?: string;
@@ -109,31 +108,29 @@ export const Contact2 = (props: IContact2Props): React.ReactElement => {
   return (
     <Section {...props as ISectionProps} className={getClassName(Contact2.displayName, props.className)}>
       <ResponsiveContainingView sizeResponsive={{ base: 10 }}>
-        <Stack directionResponsive={{ base: Direction.Horizontal }}>
-          <Stack direction={Direction.Vertical} paddingStart={EverypagePaddingSize.SectionTop} paddingEnd={EverypagePaddingSize.SectionBottom}>
-            {props.titleText && <SectionTitleText text={props.titleText} />}
-            {props.subtitleText && <Stack.Item><SectionSubtitleText text={props.subtitleText} /></Stack.Item>}
-            {props.addressLine1 && (
-              <Text variant='small-light'>
-                {props.addressLine1}
-                <br />
-                {props.addressLine2}
-              </Text>
+        <Stack direction={Direction.Horizontal}>
+          <Stack direction={Direction.Vertical} paddingStart={EverypagePaddingSize.SectionTop} paddingEnd={EverypagePaddingSize.SectionBottom} shouldAddGutters={true}>
+            {props.titleText && <Stack.Item gutterAfter={props.subtitleText ? PaddingSize.Wide : PaddingSize.Wide2}><SectionTitleText text={props.titleText} /></Stack.Item>}
+            {props.subtitleText && <Stack.Item gutterAfter={PaddingSize.Wide2}><SectionSubtitleText text={props.subtitleText} /></Stack.Item>}
+            {props.address && (
+              <Stack.Item gutterAfter={PaddingSize.Wide}>
+                <MarkdownText source={props.address} />
+              </Stack.Item>
             )}
             {props.phoneNumber && (
-              <Stack direction={Direction.Horizontal} contentAlignment={Alignment.Start} shouldAddGutters={true} defaultGutter={PaddingSize.Default}>
-                <KibaIcon _color='$colors.textLight25' iconId='ion-call-outline' />
-                <Text variant='light'>{`+ (${props.countryCode}) ${props.phoneNumber}`}</Text>
+              <Stack direction={Direction.Horizontal} contentAlignment={Alignment.Start} shouldAddGutters={true}>
+                <KibaIcon iconId='ion-call-outline' />
+                <Text>{props.phoneNumber}</Text>
               </Stack>
-            )
-            }
+            )}
             {props.email && (
-              <Stack direction={Direction.Horizontal} contentAlignment={Alignment.Start} shouldAddGutters={true} defaultGutter={PaddingSize.Default}>
-                <KibaIcon _color='$colors.textLight25' iconId='ion-mail-outline' />
-                <Text variant='light'>{props.email}</Text>
+              <Stack direction={Direction.Horizontal} contentAlignment={Alignment.Start} shouldAddGutters={true}>
+                <KibaIcon iconId='ion-mail-outline' />
+                <Text>{props.email}</Text>
               </Stack>
-            )
-            }
+            )}
+            <Stack.Item growthFactor={20} shrinkFactor={1} />
+            {props.bottomText && <Stack.Item><SectionSubtitleText text={props.bottomText} /></Stack.Item>}
           </Stack>
           <ResponsiveTextAlignmentView alignment={TextAlignment.Center}>
             <Stack direction={Direction.Vertical} paddingStart={EverypagePaddingSize.SectionTop} paddingEnd={EverypagePaddingSize.SectionBottom}>