+++ src/packages/core/src/sections/index.ts
@@ -14,7 +14,6 @@ export * from './featureIcons2';
 export * from './featureMediaHalf1';
 export * from './faq1';
 export * from './footer1';
-export * from './footer2';
 export * from './meta';
 export * from './heroAppButtonsMediaHalf1';
 export * from './heroAppDownload1';