+++ CHANGELOG.md
@@ -10,10 +10,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ### Added
 - [MINOR] Created `FeatureIcons1` section
 - [MINOR] Created `FeatureIcons2` section
+- [MINOR] Created `Contact2` section
 
 ### Changed
 - [MINOR] Update to use react v17 for all packages
 - [MINOR] Update card boxes to be full height wherever used
+- [MINOR] Created `Dialog` component for console
 
 ### Removed